Add candidate filtering and scoring, but no refinement
