+++ CMakeLists.txt
@@ -71,4 +71,3 @@ enable_testing()
 add_subdirectory(h5read)
 add_subdirectory(baseline)
 add_subdirectory(spotfinder)
-add_subdirectory(baseline_indexer)
